% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chicksdata.R
\docType{data}
\name{chicks}
\alias{chicks}
\title{Average weight of 28 chicks.}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{weight, in grams}
  \item{Chick}{identifier}
  ...
}}
\source{
\url{https://stat.ethz.ch/R-manual/R-devel/library/datasets/html/ChickWeight.html}
}
\usage{
chicks
}
\description{
A dataset containing the weights of 28 chicks.
}
\keyword{datasets}
